refactor: reorganize component folders
